pyproject.toml
[tool.poetry]
name = "aiml_newsletter_embedder"
version = "0.1.0"
description = "AI/ML Newsletter processor and semantic search engine using vector embeddings"
authors = ["Your Name <your.email@example.com>"]
readme = "README.md"
packages = [
    { include = "newsletter_processor" }
]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.109.0"
uvicorn = "^0.27.0"
apscheduler = "^3.8.1"
weaviate-client = "^3.0.0"
pydantic = "^1.8.2"
aiohttp = "^3.8.4"
async-timeout = "^4.0.2"
authlib = "^1.2.0"
langchain = "^0.0.157"
openai = "^0.27.6"
python-dotenv = "^1.0.0"
numpy = "^1.24.3"
python-multipart = "^0.0.6"

[tool.poetry.group.dev.dependencies]
pytest = "^7.3.1"
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api" 